refactor: refatorada arquitetura de pastas para adequação ao pattern mechanics/files/interface
