--- src/game/__init__.py
@@ -1,5 +1,0 @@
-from .game_flow import play_game
-from .menu_management import handle_pause_menu, render_main_menu, start_new_game, load_and_play_game
-from .game_logic import is_winner
-from .game_initialization import initialize_game
-from .game_history import show_game_history
