+++ src/interface/__init__.py
@@ -1,2 +1,6 @@
 from .display import print_board, display_message
 from .menus import show_main_menu, show_pause_menu, show_file_menu, show_game_rules, show_credits
+from .game_history import show_game_history
+from .main_menu import render_main_menu
+from .pause_menu import handle_pause_menu
+from .file_menu import handle_file_menu
