+++ src/mechanics/__init__.py
@@ -1,2 +1,6 @@
 from .move_processing import make_move
 from .user_input import handle_user_input
+from .game_logic import is_winner, count_consecutive
+from .game_initialization import initialize_game
+from .game_flow import play_game
+from .start_and_load import start_new_game, load_and_play_game
