+++ src/utils/constants.py
@@ -1,0 +1,1 @@
+HISTORY_FILE_PATH = 'saves/history.txt'
